refactor(sort-chips): improve package structure
