+++ lib/views/home_content/feed/feed_sort_option_chips.dart
@@ -1,5 +1,5 @@
 import "package:flutter/material.dart";
-import "package:proxima/views/sort_option_widgets/feed_sort_option/feed_sort_option.dart";
+import "package:proxima/models/sorting/feed_sort_option.dart";
 
 /// This widget is the list of sort options for the timeline.
 class FeedSortOptionChips extends StatelessWidget {
