+++ lib/views/home_content/feed/post_feed.dart
@@ -5,7 +5,7 @@ import "package:proxima/utils/ui/circular_value.dart";
 import "package:proxima/viewmodels/home_view_model.dart";
 import "package:proxima/views/home_content/feed/post_card/post_card.dart";
 import "package:proxima/views/navigation/routes.dart";
-import "package:proxima/views/sort_option_widgets/feed_sort_option/feed_sort_option_chips.dart";
+import "package:proxima/views/home_content/feed/feed_sort_option_chips.dart";
 
 /// This widget is the feed of the home page
 /// It contains the posts
